Pause Menu & 2D Camera
I added a pause menu where you can get to options, and quit. I also added a 2DCamera to the character and commented out the canvas moving code, because that was causing issues when moving back to the main menu. This also lets us adjust things like the camera smoothing.