Store asset and UI updates
Shows the face of the correct shop keeper. Most items in stores have correct art. The descriptions and other stats need corrections. There were issues with the some of the store assets having the same UUIDs and stuff that should be fixed now.

Now: It seems like the number of items for each slot is acting up.