Ability Cooldown Indicators
There are indicators in the bottom left for ability recharge. I linked up the dash and boombox, but the rest is listening for all of the emits when using it. It also listens for it being purchased so the icon can be enabled after we set to not visible by default